requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg[binary]==3.3.3
python-dotenv==1.0.1